nice

+++ crates/astronomical-calculator/src/tests/mod.rs
@@ -1008,7 +1008,7 @@ fn test_alos_hashachar_regression_2051_07_10() {
 
     let alos_result = calc.get_sunrise_offset_by_degrees(16.1).unwrap();
     let alos_ts = alos_result.timestamp().expect("Alos Hashachar should occur");
-    let alos_dt = Paris.timestamp_millis_opt((alos_ts * 1000) as i64).unwrap();
+    let alos_dt = Paris.timestamp_millis_opt(alos_ts * 1000).unwrap();
 
     let expected_dt = Paris
         .with_ymd_and_hms(2051, 7, 10, 3, 24, 29)
